Move server to it's own process
